// src/shared/api/generated/operation/listContentSourcesParams.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Amazing EKB Hub API
 * MVP backend API (OpenAPI-first)
 * OpenAPI spec version: 1.0.0
 */
import type { ContentSourcePlatform } from '../model/contentSourcePlatform';
import type { ContentSourceStatus } from '../model/contentSourceStatus';

export type ListContentSourcesParams = {
/**
 * Фильтр по платформе источника.
 */
platform?: ContentSourcePlatform;
/**
 * Фильтр по статусу источника.
 */
status?: ContentSourceStatus;
};